requirements.txt
attrs==21.2.0
zope.interface==5.4.0
PyYAML==6.0